+++ tests/utils/cleanup_test.go
@@ -1,7 +1,7 @@
 //go:build e2e
 // +build e2e
 
-package assets
+package utils
 
 import (
 	"fmt"
